requirements.txt
langchain==1.2.10
langchain-core==1.2.16
langchain-groq==1.1.2
langchain-community==0.4.1
langgraph==1.0.9
langsmith==0.7.7
fastapi==0.133.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
groq==0.37.1
httpx==0.28.1
requests==2.32.5
python-multipart==0.0.20
aiofiles==25.1.0
pytest==9.0.2
pytest-asyncio